+++ ios/Classes/CleverTapPlugin.h
@@ -5,6 +5,7 @@ static NSString *const kCleverTapProfileSync                = @"profileDataUpdat
 static NSString *const kCleverTapInAppNotificationDismissed = @"inAppNotificationDismissed";
 static NSString *const kCleverTapInboxDidInitialize         = @"inboxDidInitialize";
 static NSString *const kCleverTapInboxMessagesDidUpdate     = @"inboxMessagesDidUpdate";
+static NSString *const kCleverTapExperimentsDidUpdate       = @"CTExperimentsUpdated";
 
 @interface CleverTapPlugin : NSObject<FlutterPlugin>
 + (instancetype)sharedInstance;