+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGE LOG
 
+### Version 1.1.0 (January 29, 2020)
+* Adds support for Dynamic Variables & Native Display
+* Supports CleverTap Android SDK v3.6.3
+* Supports CleverTap iOS SDK v3.7.2
+
 ### Version 1.0.0 (January 20, 2020)
 * Initial Release.
 * Supports CleverTap Android SDK v3.6.1